requirements.txt
pyqt6
numpy
pillow
pyyaml
